IPE FLTv4: writeThreshold now use int rather than short; changed limits for postTriggTime. -tb-
git-svn-id: svn://orca.physics.unc.edu/Orca@2848 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/IPE/IpeV4 FLT/ORIpeV4FLTModel.h
@@ -307,8 +307,8 @@
 
 /** Print result of hardware statistics for all channels */
 - (void) printStatistics; // ak, 7.10.07
-- (void) writeThreshold:(int)i value:(unsigned short)aValue;
-- (unsigned short) readThreshold:(int)i;
+- (void) writeThreshold:(int)i value:(unsigned int)aValue;
+- (unsigned int) readThreshold:(int)i;
 - (void) writeGain:(int)i value:(unsigned short)aValue;
 - (unsigned short) readGain:(int)i;
 - (void) writeTriggerControl;
